fixed additional occurrences of size_t / int warnings of AppVeyor:
..\crypto\packet.c(277) + (281)
..\crypto\param_build.c(92) + (232) + (250) + (268) + (284) + (300) + (332)
..\crypto\params.c(940)
..\crypto\params_from_text.c(131)

+++ crypto/bio/bio_lib.c
@@ -348,7 +348,7 @@ static int bio_write_intern(BIO *b, const void *data, size_t dlen,
     return ret;
 }
 
-int BIO_write(BIO *b, const void *data, int dlen)
+int BIO_write(BIO *b, const void *data, size_t dlen)
 {
     size_t written;
     int ret;
@@ -356,7 +356,7 @@ int BIO_write(BIO *b, const void *data, int dlen)
     if (dlen < 0)
         return 0;
 
-    ret = bio_write_intern(b, data, (size_t)dlen, &written);
+    ret = bio_write_intern(b, data, dlen, &written);
 
     if (ret > 0) {
         /* *written should always be <= dlen */
